Do not construct values from other keys values within a source
We will later do that in a "derived"/"convert" source,
after a "selector" source,
which chooses the best value for each property first.
The "selector" source will probably also require
a per-value "faith" value.

This way, we should be less likely ending up
with values that do not fit to each other,
like a GitHub.com clone URL and a BitBucket.org issues URL.